server/requirements.txt
flask==2.0.3
werkzeug==2.0.3
pillow==9.4.0
requests==2.28.2